@@ -50,7 +50,7 @@ function session_begin($user_ip): void
  */
 function session(): void
 {
-    global $user_ip, $cookie_id, $server_config;
+    global $user_ip, $cookie_id, $server_config, $log;
     global $_COOKIE;
     $Sessions_Model = new Sessions_Model();
 
@@ -60,32 +60,46 @@ function session(): void
 
     //Purge des sessions expirées
     if ($server_config["session_time"] != 0) {
+        $log->debug("Cleaning expired sessions");
         $Sessions_Model->clean_expired_sessions();
     }
 
     //Récupération de l'id de session si cookie présent
     if (isset($_COOKIE[$cookie_name])) {
         $cookie_id = $_COOKIE[$cookie_name];
+        $log->debug("Session cookie found: " . substr($cookie_id, 0, 8) . "...");
 
         //Vérification de la validité de le session
         if (!$Sessions_Model->is_valid_session_id($cookie_id, $user_ip)) {
+            $log->warning("Invalid session for cookie_id: " . $cookie_id . " and user_ip: " . $user_ip);
+
             if (isset($server_config["disable_ip_check"]) && $server_config["disable_ip_check"] == 1) {
                 //Mise à jour de l'adresse ip de session si le contrôle des ip est désactivé
+                $log->info("Attempting to update IP address - IP check disabled");
+
                 if (!$Sessions_Model->update_session_public_ip($cookie_id, $user_ip)) {
+                    $log->warning("Failed to update session IP address");
                     $cookie_id = "";
+                } else {
+                    $log->info("Session IP address updated successfully");
                 }
             } else {
                 $cookie_id = "";
             }
+        } else {
+            $log->debug("Valid session found");
         }
+    } else {
+        $log->debug("No session cookie found");
     }
     if ($cookie_id == "") {
+        $log->info("Creating new session for IP address: " . $user_ip);
         session_begin($user_ip);
     } else {
         $cookie_expire = time() + $cookie_time * 60;
+        $log->debug("Updating session expiration time");
         $Sessions_Model->update_session_expiration_time($cookie_id, $cookie_expire);
     }
-
     session_set_user_data($cookie_id);
 }
 
@@ -96,16 +110,21 @@ function session(): void
  */
 function session_set_user_id($user_id, $lastvisit = 0): void
 {
-    global $user_ip, $cookie_id, $server_config;
+    global $user_ip, $cookie_id, $server_config,$log;
     $Sessions_Model = new Sessions_Model();
 
+    $log->info("Updating session for user ID: " . $user_id);
+
     if (isset($server_config["disable_ip_check"]) && $server_config["disable_ip_check"] != 1) {
+        $log->debug("Updating session with IP verification enabled");
         $Sessions_Model->update_session($user_id, $lastvisit, $cookie_id, $user_ip);
     } else {
+        $log->debug("Updating session without IP verification");
         $Sessions_Model->update_session($user_id, $lastvisit, $cookie_id);
     }
-
+    $log->debug("Retrieving user data after session update");
     session_set_user_data($cookie_id);
+    $log->info("User session ID: " . $user_id . " updated successfully");
 }
 
 /**
@@ -115,14 +134,23 @@ function session_set_user_id($user_id, $lastvisit = 0): void
  */
 function session_set_user_data(string $cookie_id)
 {
-    global $user_ip, $user_data, $user_auth, $user_token;
+    global $user_ip, $user_data, $user_auth, $user_token,$log;
 
+    $log->debug("Retrieving user data for cookie_id: " . substr($cookie_id, 0, 8) . "...");
     $user_data = (new Sessions_Model())->select_user_data_session($cookie_id, $user_ip);
 
     if (!$user_data) {
+        $log->warning("No user data found for cookie_id: " . substr($cookie_id, 0, 8) . "...");
         unset($user_data);
         unset($user_auth);
         unset($user_token);
+    } else {
+        if (isset($user_data["id"])) {
+            $log->info("User identified - ID: " . $user_data["id"] . ", username: " . $user_data["name"]);
+        } else {
+            $log->info("Unidentified visitor");
+        }
+
     }
 }
 
@@ -163,18 +191,18 @@ function session_whois_online()
 
     $tOnline = (new Sessions_Model())->who_is_online();
 
-    $guests = array();
-    $members = array();
+    $guests = [];
+    $members = [];
     foreach ($tOnline as $online) {
         $time_lastactivity = $online["session_expire"] - $cookie_time * 60;
         $session_ip = decode_ip($online["session_ip"]);
 
-        if (is_null($online["user_name"])) {
+        if (is_null($online["name"])) {
             $username = "Visiteur non identifié";
-            $guests[] = array("user" => $username, "time_start" => $online["session_start"], "time_lastactivity" => $time_lastactivity, "ip" => $session_ip, "ogs" => 0);
+            $guests[] = array("user" => $username, "time_start" => $online["session_start"], "time_lastactivity" => $time_lastactivity, "ip" => $session_ip, "session_type" => 0);
         } else {
-            $username = $online["user_name"];
-            $members[] = array("user" => $username, "time_start" => $online["session_start"], "time_lastactivity" => $time_lastactivity, "ip" => $session_ip, "ogs" => $online["session_ogs"]);
+            $username = $online["name"];
+            $members[] = array("user" => $username, "time_start" => $online["session_start"], "time_lastactivity" => $time_lastactivity, "ip" => $session_ip, "session_type" => $online["session_type"]);
         }
     }
     return array_merge($members, $guests);
