@@ -39,7 +39,7 @@ $lang['GALAXY_RANK_MILITARY'] = "Vojska";
 $lang['GALAXY_RANK_MILITARY_BUILT'] = "Izgrađeno vojske";
 $lang['GALAXY_RANK_MILITARY_LOST'] = "Izgubljeno vojske";
 $lang['GALAXY_RANK_MILITARY_DESTROYED'] = "Uništeno vojske";
-$lang['GALAXY_RANK_MILITARY_HONNOR'] = "Čast";
+$lang['GALAXY_RANK_MILITARY_HONOR'] = "Čast";
 $lang['GALAXY_MEMBERS'] = "član(ovi)";
 $lang['GALAXY_SEE_DETAILS'] = "Pogledaj detaljnije";
 $lang['GALAXY_CR'] = "CR";
