@@ -39,7 +39,7 @@ $lang['GALAXY_RANK_MILITARY'] = "Military";
 $lang['GALAXY_RANK_MILITARY_BUILT'] = "Militaire built";
 $lang['GALAXY_RANK_MILITARY_LOST'] = "Militaire lost";
 $lang['GALAXY_RANK_MILITARY_DESTROYED'] = "Militaire destroyed";
-$lang['GALAXY_RANK_MILITARY_HONNOR'] = "Honnor";
+$lang['GALAXY_RANK_MILITARY_HONOR'] = "Honor";
 $lang['GALAXY_MEMBERS'] = "member(s)";
 $lang['GALAXY_SEE_DETAILS'] = "See details";
 $lang['GALAXY_CR'] = "CR";
