@@ -35,7 +35,7 @@ $lang['GALAXY_RANK_MILITARY'] = "Militaire";
 $lang['GALAXY_RANK_MILITARY_BUILT'] = "Militaire construit";
 $lang['GALAXY_RANK_MILITARY_LOST'] = "Perte militaire";
 $lang['GALAXY_RANK_MILITARY_DESTROYED'] = "Destruction";
-$lang['GALAXY_RANK_MILITARY_HONNOR'] = "Honneur";
+$lang['GALAXY_RANK_MILITARY_HONOR'] = "Honneur";
 $lang['GALAXY_MEMBERS'] = "membre(s)";
 $lang['GALAXY_SEE_DETAILS'] = "Voir détails";
 $lang['GALAXY_CR'] = "RC";
