@@ -15,25 +15,40 @@ if (!defined('IN_SPYOGAME')) {
     die("Hacking attempt");
 }
 
-if ($user_data["user_admin"] != 1 && $user_data["user_coadmin"] != 1) {
+if ($user_data["admin"] != 1 && $user_data["coadmin"] != 1) {
     redirection("index.php?action=message&amp;id_message=forbidden&amp;info");
 }
 
+// Niveaux de log Monolog avec leur valeur numérique
+$monolog_levels = [
+    'EMERGENCY' => 600,
+    'ALERT'     => 550,
+    'CRITICAL'  => 500,
+    'ERROR'     => 400,
+    'WARNING'   => 300,
+    'NOTICE'    => 250,
+    'INFO'      => 200,
+    'DEBUG'     => 100,
+];
+
+//Définition du niveau de log sélectionné
+$log_level_filter = isset($pub_level) && isset($monolog_levels[$pub_level]) ? $pub_level : 'INFO';
+
 //Définition de la date sélectionnée
 if (!isset($pub_show)) {
-    $show = date("y~m~d");
+    $show = date("Y~m~d");
 } else {
     $show = $pub_show;
 }
 
 if (sizeof(explode("~", $show)) != 3) {
-    $show = date("y~m~d");
+    $show = date("Y~m~d");
 }
 list($show_year, $show_month, $show_day) = explode("~", $show);
 if (!checkdate($show_month, $show_day, $show_year)) {
-    list($show_year, $show_month, $show_day) = explode("~", date("y~m~d"));
+    list($show_year, $show_month, $show_day) = explode("~", date("Y~m~d"));
 }
-$show = $show_year . $show_month . $show_day;
+$show = "$show_year-$show_month-$show_day";
 
 if (!isset($pub_typelog)) {
     $typelog = "log";
@@ -48,20 +63,20 @@ if ($typelog != "log" && $typelog != "sql") {
     $typelog = "log";
 }
 if ($typelog == "log") {
-    $file = $typelog . "_" . $show . ".log";
+    $file = "OGSpy-" . $show . ".log";
 }
 if ($typelog == "sql") {
-    $file = $typelog . "_" . $show . ".sql";
+    $file = "OGSpy-sql-" . $show . ".log";
 }
 
 //Récupération du log
 $dir = $show;
-$file = PATH_LOG . $dir . "/" . $file;
+$file = PATH_LOG . "/" . $file;
 
 if (file_exists($file)) {
-    $log = file($file);
+    $logFileName = file($file);
 } else {
-    $log = array($lang['ADMIN_LOGS_NOLOGS']);
+    $logFileName = array($lang['ADMIN_LOGS_NOLOGS']);
 }
 
 echo "<!--<a>" . $lang['ADMIN_LOGS_SELECTED_DATE'] .  date("d F o", mktime(0, 0, 0, $show_month, $show_day, $show_year)) . "</a>-->";
@@ -79,37 +94,26 @@ echo "<!--<a>" . $lang['ADMIN_LOGS_SELECTED_DATE'] .  date("d F o", mktime(0, 0,
          <?php $date = mktime(0, 0, 0, date("n"), 1) - 60 * 60 * 24 * 365; ?>
     <tr> <!-- Mois -->
     <?php for ($i = 0; $i < 12; $i++)  :?>
-        <?php $date += 60 * 60 * 24 * 31;?>
-        <?php $show = date("y~m", $date) . "~" . $show_day;?>
+        <?php
+        $date += 60 * 60 * 24 * 31;
+        $show_month_loop = date("m", $date);
+        $show_year_loop = date("Y", $date);
+        $show_param = $show_year_loop . "~" . $show_month_loop . "~" . $show_day;
 
-        <?php if ($show == $show_year . "~" . $show_month . "~" . $show_day) :  ?>
-            <?php if (log_check_exist(date("ym", $date))) :?>
-            <td>
-                <span class="og-success"><?= date("F o", $date) ?></span>
-            </td>
-             <td>
-                 <input class="og-button og-button-image  og-button-warning" type='image' src='images/save.png' onclick="window.location = 'index.php?action=extractor&amp;date=<?= date("Fo", $date)?>'" title='<?= $lang['ADMIN_LOGS_DOWNLOAD'] . date("F o", $date) ?>'>
-                 <input  class="og-button og-button-image  og-button-danger" type='image' src='images/drop.png' onclick="window.location = 'index.php?action=remove&amp;date=<?php date("ym", $date) . $show_day?>&directory=TRUE'" title='<?= $lang['ADMIN_LOGS_DELETE'] . date("F o", $date) ?>'>
-            <?php else : ?>
-                <td colspan='2'>
-                    <?= date("F o", $date) ?>
-                </td>
-            <?php endif ;?>
-           </td>
-        <?php else :?>
-            <?php if (log_check_exist(date("ym", $date))) :?>
-                <td onclick="window.location = 'index.php?action=administration&amp;subaction=viewer&amp;show=<?= $show ?>&amp;typelog=<?= $typelog ?>';">
-                    <?= date("F o", $date) ?>
-                </td>
-                <td>
-                    <input class="og-button og-button-image  og-button-warning" type='image' src='images/save.png' onclick="window.location = 'index.php?action=extractor&amp;date=<?= date("ym", $date) ?>'" title='<?= $lang['ADMIN_LOGS_DOWNLOAD'] . date("F o", $date) ?>'>
-                </td>
-            <?php else : ?>
-               <td colspan='2' onclick="window.location = 'index.php?action=administration&amp;subaction=viewer&amp;show=<?php $show . "&amp;typelog=" . $typelog;?>';">
-                   <?= date("F o", $date) ?>
-               </td>
-            <?php endif ;?>
-        <?php endif;?>
+        $log_exists = log_check_exist(date("Ym", $date));
+        $is_selected_month = ($show_year_loop == $show_year && $show_month_loop == $show_month);
+
+        $cell_class = '';
+        if ($log_exists) {
+            $cell_class = 'og-success';
+        }
+        if ($is_selected_month) {
+            $cell_class .= ' og-highlight';
+        }
+        ?>
+        <td colspan="2" class="<?= $cell_class ?>" style="cursor: pointer;" onclick="window.location = 'index.php?action=administration&amp;subaction=viewer&amp;show=<?= $show_param ?>&amp;typelog=<?= $typelog ?>&level=<?= $log_level_filter ?>';">
+            <?= date("F Y", $date) ?>
+        </td>
         <?php if ($i == 5) :?>
             </tr>
             <tr>
@@ -122,61 +126,53 @@ echo "<!--<a>" . $lang['ADMIN_LOGS_SELECTED_DATE'] .  date("d F o", mktime(0, 0,
 <table class="og-table og-full-table">
     <thead>
         <tr>
-            <th  colspan="20">
+            <th colspan="10">
                 <?= ($lang['ADMIN_LOGS_SELECT_DAY']) ?>
              </th>
         </tr>
     </thead>
     <tbody>
-         <?php $max_day = (intval($show_month) != date("m")) ? date("t", mktime(0, 0, 0, $show_month, 1, $show_year)) : date("d") ;?>
+        <?php
+        $max_day = date("t", mktime(0, 0, 0, $show_month, 1, $show_year));
+        $current_day_in_loop = 1;
+        ?>
     <tr> <!-- Jour -->
-   <?php  for ($i = 1; $i <= $max_day; $i++) :?>
-        <?php $day = $i;?>
-        <?php if ($i < 10) :?>
-            <?php $day = "0" . $i;?>
-        <?php endif;?>
-        <?php $show = $show_year . "~" . $show_month . "~" . $day;?>
-        <?php $date = mktime(0, 0, 0, $show_month, $day, $show_year);?>
+   <?php for ($i = 1; $i <= $max_day; $i++) :?>
+        <?php
+        $day = str_pad($i, 2, '0', STR_PAD_LEFT);
+        $current_show_date_param = $show_year . "~" . $show_month . "~" . $day;
+        $log_exists = log_check_exist($show_year . $show_month . $day);
 
-        <?php if ($show == $show_year . "~" . $show_month . "~" . $show_day) :?>
-            <?php if (log_check_exist($show_year . $show_month . $day)): ?>
-        <td>
-            <span class="og-success"><?= $day ?></span>
+        // La date sélectionnée via l'URL est au format Y~m~d
+        $selected_date_from_url = $pub_show ?? date("Y~m~d");
+
+        $is_selected_day = ($current_show_date_param == $selected_date_from_url);
+
+        $cell_class = '';
+        if ($log_exists) {
+            $cell_class = 'og-success'; // Vert pour les jours avec logs
+        }
+        if ($is_selected_day) {
+            $cell_class .= ' og-highlight'; // Jaune/Orange pour le jour sélectionné
+        }
+        ?>
+        <td class="<?= $cell_class ?>" style="cursor: pointer;" onclick="window.location='index.php?action=administration&subaction=viewer&show=<?= $current_show_date_param ?>&typelog=<?= $typelog ?>&level=<?= $log_level_filter ?>';">
+            <?= $day ?>
         </td>
-                <td>
-                    <input class="og-button og-button-image  og-button-warning" type='image' src='images/save.png' onclick="window.location = 'index.php?action=extractor&amp;date=<?= $show_year . $show_month . $day ?>'" title='<?= $lang['ADMIN_LOGS_DOWNLOAD'] . date("d F o", $date) ?>'>
-                    <input class="og-button og-button-image  og-button-danger" type='image' src='images/drop.png' onclick="window.location = 'index.php?action=remove&amp;date=<?= $show_year . $show_month . $day ?>'" title='<?= $lang['ADMIN_LOGS_DELETE'] . date("d F o", $date) ?>'>
-                </td>
-            <?php else : ?>
-                <td colspan='2'>
-                    <?= $day ?>
-                </td>
-            <?php endif; ?>
-
-        <?php else :?>
-            <?php if (log_check_exist($show_year . $show_month . $day)) :?>
-                <td onclick="window.location='index.php?action=administration&amp;subaction=viewer&amp;show=<?= $show ?>&amp;typelog=<?= $typelog ?>';">
-                    <?= $day ?>
-                 </td>
-                 <td>
-                     <input class="og-button og-button-image  og-button-warning" type='image' src='images/save.png' onclick="window.location = 'index.php?action=extractor&amp;date=<?= $show_year . $show_month . $day ?>'" title='<?= $lang['ADMIN_LOGS_DOWNLOAD'] . date("d F o G:i", $date)?>'>
-            <?php else : ?>
-                <td colspan='2' onclick="window.location='index.php?action=administration&amp;subaction=viewer&amp;show=<?= $show ?>&amp;typelog=<?= $typelog ?>';">
-                    <?= $day ?>
-                </td>
-            <?php endif ;?>
-        <?php endif ;?>
-        <?php if ($i % 10 == 0) :?>
+        <?php if ($i % 10 == 0 && $i < $max_day) :?>
             </tr>
             <tr>
         <?php endif ;?>
     <?php endfor ;?>
-    <?php $j = 1; /* $j ?*/?>
-    <?php while (($i - 1) % 10 != 0) :?>
-       <td colspan='2'></td>
-        <?php $i++;?>
-    <?php endwhile;?>
-
+    <?php
+    // Complète la dernière ligne avec des cellules vides si nécessaire
+    $remaining_cells = 10 - ($max_day % 10);
+    if ($remaining_cells < 10) {
+        for ($j = 0; $j < $remaining_cells; $j++) {
+            echo "<td></td>";
+        }
+    }
+    ?>
     </tr>
     </tbody>
 </table>
@@ -200,6 +196,28 @@ echo "<!--<a>" . $lang['ADMIN_LOGS_SELECTED_DATE'] .  date("d F o", mktime(0, 0,
 
 </div>
 
+<div style="margin-bottom: 10px;">
+    <form action="index.php" method="get" style="display: inline;">
+        <input type="hidden" name="action" value="administration">
+        <input type="hidden" name="subaction" value="viewer">
+        <input type="hidden" name="show" value="<?= $show ?>">
+        <input type="hidden" name="typelog" value="<?= $typelog ?>">
+        <label for="level">Niveau de log minimum :</label>
+        <select name="level" id="level" onchange="this.form.submit()">
+            <?php foreach ($monolog_levels as $level_name => $level_value) : ?>
+                <option value="<?= $level_name ?>" <?= ($log_level_filter == $level_name) ? 'selected' : '' ?>>
+                    <?= $level_name ?>
+                </option>
+            <?php endforeach; ?>
+        </select>
+    </form>
+    <?php if (file_exists($file)) : ?>
+        <a href="logs/<?= basename($file) ?>" download title="Télécharger le fichier de log" style="display: inline-block; vertical-align: middle; margin-left: 10px;">
+            <img src="images/save.png" alt="Télécharger le fichier de log" />
+        </a>
+    <?php endif; ?>
+</div>
+
 <span class="og-alert"><?= ($lang['ADMIN_LOGS_SEE_TRANSACTIONALS']) ?></span>
 
 <table class="og-table og-full-table og-table-log">
@@ -215,21 +233,59 @@ echo "<!--<a>" . $lang['ADMIN_LOGS_SELECTED_DATE'] .  date("d F o", mktime(0, 0,
     </thead>
     <tbody>
           <tr>
-            <td class="tdvalue" >
+              <td class="tdvalue" >
+                  <?php
+                  // Niveaux Monolog et leurs classes CSS associées
+                  $log_levels = [
+                      'EMERGENCY' => 'log-emergency',
+                      'ALERT'     => 'log-alert',
+                      'CRITICAL'  => 'log-critical',
+                      'ERROR'     => 'log-error',
+                      'WARNING'   => 'log-warning',
+                      'NOTICE'    => 'log-notice',
+                      'INFO'      => 'log-info',
+                      'DEBUG'     => 'log-debug',
+                  ];
 
- <?php
-            end($log);
-            while ($line = current($log)) {
+                  $min_level_value = $monolog_levels[$log_level_filter];
 
-                $line = trim(nl2br(htmlspecialchars($line)));
-                $line = preg_replace("#/\*(.*)\*/#", "<span style=\"color: orange; \">$1 : </span>", $line);
+                  foreach (array_reverse($logFileName) as $line) {
+                      $line = rtrim($line, "\r\n");
+                      if (empty($line)) {
+                          continue;
+                      }
 
-                echo $line;
-                prev($log);
-            }
-            ?>
-            </td>
+                      $current_line_level = null;
+                      foreach ($monolog_levels as $level_name => $level_value) {
+                          if (preg_match('/OGSpy(SQL)?\.' . $level_name . ':/', $line)) {
+                              $current_line_level = $level_value;
+                              break;
+                          }
+                      }
+
+                      // Si le niveau de log n'est pas trouvé ou est inférieur au filtre, on passe à la ligne suivante
+                      if ($current_line_level === null || $current_line_level < $min_level_value) {
+                          continue;
+                      }
+
+                      $applied_class = 'log-line'; // Classe par défaut
+
+                      // Cherche le niveau de log dans l'entrée
+                      foreach ($log_levels as $level => $class) {
+                          if (preg_match('/OGSpy(SQL)?\.' . $level . ':/', $line)) {
+                              $applied_class .= ' ' . $class;
+                              break; // On a trouvé le niveau, on arrête la recherche
+                          }
+                      }
+
+                      // Sécurise l'entrée pour l'affichage HTML
+                      $formatted_entry = htmlspecialchars($line, ENT_QUOTES, 'UTF-8');
+
+                      // Affiche l'entrée formatée.
+                      echo '<span class="' . $applied_class . '">' . $formatted_entry . '</span>';
+                  }
+                  ?>
+              </td>
         </tr>
         </tbody>
 </table>
-
