@@ -36,13 +36,13 @@ if (!defined('IN_SPYOGAME')) {
                 <li class='menuitem-profile'>
                     <a href='index.php?action=profile' ><?php echo ($lang['MENU_PROFILE']); ?></a>
                 </li>
-                <?php if ($user_data["user_admin"] == 1 || $user_data["user_coadmin"] == 1 || $user_data["management_user"] == 1) : ?> 
+                <?php if ($user_data["admin"] == 1 || $user_data["coadmin"] == 1 || $user_data["management_user"] == 1) : ?>
                     <li class='menuitem-administration'>
                         <a href='index.php?action=administration' ><?php echo ($lang['MENU_ADMIN']); ?></a>
                     </li>
                 <?php endif; ?>
                 <!-- partie admin prise en charge dans page admin
-                <?php if ($user_data["user_admin"] == 1 || $user_data["user_coadmin"] == 1) : ?>
+                <?php if ($user_data["admin"] == 1 || $user_data["coadmin"] == 1) : ?>
                     <li class='menuitem-admin-infoserver'>
                         <a href='index.php?action=administration&amp;subaction=infoserver' ><?php echo ($lang['ADMIN_TITLE_GENERAL_INFO']); ?></a>
                     </li>
@@ -59,7 +59,7 @@ if (!defined('IN_SPYOGAME')) {
                 <li class='menuitem-admin-group'>
                     <a href='index.php?action=administration&amp;subaction=group' ><?php echo ($lang['ADMIN_TITLE_GROUP_CONF']); ?></a>
                 </li>
-                <?php if ($user_data["user_admin"] == 1 || $user_data["user_coadmin"] == 1) : ?>
+                <?php if ($user_data["admin"] == 1 || $user_data["coadmin"] == 1) : ?>
                     <li class='menuitem-admin-viewer'>
                         <a href='index.php?action=administration&amp;subaction=viewer' ><?php echo ($lang['ADMIN_TITLE_LOGS_CONF']); ?></a>
                     </li>
@@ -74,7 +74,7 @@ if (!defined('IN_SPYOGAME')) {
                 <li class='menuitem-logout'>
                     <a href='index.php?action=logout'><?php echo $lang['MENU_LOGOUT']; ?></a>
                 </li>
-            </ul>            
+            </ul>
         </li>
 
 
@@ -113,7 +113,7 @@ if (!defined('IN_SPYOGAME')) {
                 <li class='menuitem-ogspy-statistic'>
                     <a href='index.php?action=statistic' ><?php echo ($lang['MENU_UPDATE_STATUS']); ?></a>
                 </li>
-            </ul>            
+            </ul>
         </li>
 
         <li class='menuitem-mod'>
@@ -133,9 +133,9 @@ if (!defined('IN_SPYOGAME')) {
                     <?php endif; ?>
                 <?php endforeach; ?>
                 <!-- mod admin -->
-            </ul>            
+            </ul>
             <ul class="sub_menu_navigate sub_menu_navigate_admin">
-                <?php if ($user_data["user_admin"] == 1 || $user_data["user_coadmin"] == 1) : ?>
+                <?php if ($user_data["admin"] == 1 || $user_data["coadmin"] == 1) : ?>
                     <?php foreach ($tMods as $mod) : ?>
                         <?php if ($mod['admin_only'] == 1) : ?>
                             <li class='menusubitem-mod-admin menusubitem-mod-<?php echo $mod['action']; ?>'>
@@ -165,7 +165,4 @@ if (!defined('IN_SPYOGAME')) {
         </li>
 
     </ul>
-</div>    
-
-
-
+</div>
