@@ -17,7 +17,7 @@ if (!defined('IN_SPYOGAME')) {
 <html lang="<?= $lang['HEAD_LANGUAGE']; ?>">
 <head>
     <meta http-equiv="Content-Type" content="application/xhtml+xml; charset=UTF-8"/>
-    <title><?php echo $server_config["servername"] . " - OGSpy " . $server_config["version"]; ?></title>
+    <title><?php echo (isset($server_config["servername"]) ? $server_config["servername"] : "OGSpy") . " - OGSpy " . (isset($server_config["version"]) ? $server_config["version"] : ""); ?></title>
     <link rel="stylesheet" type="text/css" href="./skin/OGSpy_skin/formateredesign.css" />
     <link rel="shortcut icon" type="image/x-icon" href="favicon.ico"/>
     <link rel="icon" type="image/icon" href="favicon.ico"/>
@@ -25,4 +25,3 @@ if (!defined('IN_SPYOGAME')) {
 <body>
 
     <section id="content"> <!-- Contenu principal Attention, fermeture dans le footer / compat legacy -->
-
