@@ -55,7 +55,7 @@ require_once('views/page_header_2.php');
             <?php else : ?>
                 <input class="og-button og-button-warning" type='button' value='<?php echo $lang['REPORT_REMOVEFROMFAV']; ?>' onclick="<?php echo $string_delfavorites ?>">
             <?php endif; ?>
-            <?php if ($user_data['user_admin'] == 1 || $user_data['user_coadmin'] == 1) : ?>
+            <?php if ($user_data['admin'] == 1 || $user_data['coadmin'] == 1) : ?>
                 <input class="og-button og-button-danger" type='button' value='<?php echo $lang['REPORT_DELETE']; ?>' onclick="<?php echo $string_delspy; ?>">
             <?php endif; ?>
             <?php echo $v['data']; ?><br /><br />
